moved weights from int to float

+++ DVSTracker/filter.h
@@ -8,19 +8,19 @@ class Filter
 public:
     Filter(int size, float standardDeviation, int mapW, int mapH);
     ~Filter();
-    Map<int>* smoothen(Map<int> *buffer);
+    Map<float>* smoothen(Map<float> *buffer);
 
     static const float PI;
 
 private:
     //operations on kernel
-    int kernelGet(int x, int y);
+    float kernelGet(int x, int y);
     void kernelSet(int x, int y, float value);
     void generateKernel();
     float convolute();
-    bool outOfBounds(Map<int> *buffer, int x, int y);
+    bool outOfBounds(Map<float> *buffer, int x, int y);
 
-    Map<int> *filteredMap;
+    Map<float> *filteredMap;
     int kernelSize;
     float sd;
     float *kernel;
