+++ DVSTracker/localmaximum.cpp
@@ -1,13 +1,13 @@
 #include "localmaximum.h"
 
-LocalMaximum::LocalMaximum(int posX, int posY, int w)
+LocalMaximum::LocalMaximum(int posX, int posY, float w)
 {
     x = posX;
     y = posY;
     weight = w;
 }
 
-void LocalMaximum::set(int posX, int posY, int w){
+void LocalMaximum::set(int posX, int posY, float w){
     x = posX;
     y = posY;
     weight = w;
